requirements.txt
pytest-playwright
numpy
pandas
pydantic